+++ assets/js/constants.ts
@@ -1,0 +1,9 @@
+// FIXME: Move all UI style variables here
+export const COLORS = {
+    "lightgray": "#e5e5e5"
+}
+
+export const FONTSIZES = {
+    "small": 10,
+    "medium": 18
+}