+++ assets/js/track/annotation.ts
@@ -26,7 +26,28 @@ function stringToHash(string) {
 }
 
 export class AnnotationTrack extends BaseAnnotationTrack {
-  constructor(x, width, near, far, genomeBuild, defaultAnnotation) {
+  readonly featureHeight: number = 18;
+  readonly arrowThickness: number = 2;
+  readonly apiEntrypoint: string = "get-annotation-data";
+
+  genomeBuild: number;
+  sourceList: HTMLSelectElement;
+  additionalQueryParams: { source: string };
+  numRenderedElements: number;
+  heightOrderRecord: {
+    latestHeight: number;
+    latestNameEnd: number;
+    latestTrackEnd: number;
+  };
+
+  constructor(
+    x: number,
+    width: number,
+    near: number,
+    far: number,
+    genomeBuild: number,
+    defaultAnnotation: string
+  ) {
     // Dimensions of track canvas
     const visibleHeight = 300; // Visible height for expanded canvas, overflows for scroll
     const minHeight = 35; // Minimized height
@@ -39,8 +60,6 @@ export class AnnotationTrack extends BaseAnnotationTrack {
     this.contentCanvas = document.getElementById("annotation-content");
     this.trackTitle = document.getElementById("annotation-titles");
     this.trackContainer = document.getElementById("annotation-track-container");
-    this.featureHeight = 18;
-    this.arrowThickness = 2;
 
     // Setup html objects now that we have gotten the canvas and div elements
     this.setupHTML(x + 1);
@@ -49,19 +68,21 @@ export class AnnotationTrack extends BaseAnnotationTrack {
     this.genomeBuild = genomeBuild;
 
     // Setup annotation list
-    this.sourceList = document.getElementById("source-list");
+    this.sourceList = document.getElementById(
+      "source-list"
+    ) as HTMLSelectElement;
     this.sourceList.addEventListener("change", () => {
       this.expanded = false;
       this.additionalQueryParams = { source: this.sourceList.value };
-      const region = parseRegionDesignation(
-        document.getElementById("region-field").value,
-      );
+      const regionField = document.getElementById(
+        "region-field"
+      ) as HTMLInputElement;
+      const region = parseRegionDesignation(regionField.value);
       this.drawTrack({ forceRedraw: true, ...region });
     });
     this.annotSourceList(defaultAnnotation);
 
     // GENS api parameters
-    this.apiEntrypoint = "get-annotation-data";
     this.additionalQueryParams = { source: defaultAnnotation };
 
     this.maxResolution = 6; // define other max resolution
@@ -70,12 +91,13 @@ export class AnnotationTrack extends BaseAnnotationTrack {
   }
 
   // Fills the list with source files
-  annotSourceList(defaultAnnotation) {
+  annotSourceList(defaultAnnotation: string) {
     get("get-annotation-sources", { genome_build: this.genomeBuild })
       .then((result) => {
         if (result.sources.length > 0) {
           this.sourceList.style.visibility = "visible";
         }
+
         for (const fileName of result.sources) {
           // Add annotation file name to list
           const opt = document.createElement("option");
@@ -84,20 +106,22 @@ export class AnnotationTrack extends BaseAnnotationTrack {
 
           // Set mimisbrunnr as default file
           if (fileName.match(defaultAnnotation)) {
-            opt.setAttribute("selected", true);
+            opt.selected = true;
           }
           this.sourceList.appendChild(opt);
         }
       })
       .then(() => {
-        const region = parseRegionDesignation(
-          document.getElementById("region-field").value,
-        );
+        const regionField = document.getElementById(
+          "region-field"
+        ) as HTMLInputElement;
+        const region = parseRegionDesignation(regionField.value);
         this.drawTrack({ ...region });
       });
   }
 
   // Draws annotations in given range
+  // @ts-expect-error - FIXME, resolve the type errors here
   async drawOffScreenTrack({ startPos, endPos, maxHeightOrder, data }) {
     const textSize = 10;
 
@@ -122,7 +146,7 @@ export class AnnotationTrack extends BaseAnnotationTrack {
         isElementOverlapping(annot, {
           start: startPos,
           end: endPos,
-        }),
+        })
       );
     }
     // dont show tracks with no data in them
@@ -170,7 +194,7 @@ export class AnnotationTrack extends BaseAnnotationTrack {
         y2: canvasYPos + this.featureHeight / 2,
         features: [],
         isDisplayed: false,
-      };
+      } as DisplayElement;
       // Draw box for annotation
       drawRect({
         ctx: this.drawCtx,